+++ genmonlib/controller.py
@@ -1508,7 +1508,6 @@ class GeneratorController(MySupport):
     def EditMaintLogRow(self, InputString):
 
         # { index : {maint log entry}}
-        self.LogError("Edit: " + InputString)
         ValidInput = False
         EntryString = InputString
         if EntryString == None or not len(EntryString):